prog

+++ enginetest/queries/create_table_queries.go
@@ -206,13 +206,13 @@ var CreateTableQueries = []WriteQueryTest{
 		WriteQuery:          `CREATE TABLE t1 as select s, sum(i) from mytable group by s`,
 		ExpectedWriteResult: []sql.Row{{types.NewOkResult(3)}},
 		SelectQuery:         `select * from t1 order by s`, // other column is named `SUM(mytable.i)`
-		ExpectedSelect:      []sql.Row{{"first row", 1}, {"second row", 2}, {"third row", 3}},
+		ExpectedSelect:      []sql.Row{{"first row", float64(1)}, {"second row", float64(2)}, {"third row", float64(3)}},
 	},
 	{
 		WriteQuery:          `CREATE TABLE t1 as select s, sum(i) from mytable group by s having sum(i) > 2`,
 		ExpectedWriteResult: []sql.Row{{types.NewOkResult(1)}},
 		SelectQuery:         "select * from t1",
-		ExpectedSelect:      []sql.Row{{"third row", 3}},
+		ExpectedSelect:      []sql.Row{{"third row", float64(3)}},
 	},
 	{
 		WriteQuery:          `CREATE TABLE t1 as select s, i from mytable order by s limit 1`,
